Finished Part 3
Messed around with gdb and have a better understanding of how it and the
project work.

Basically, in `main` of `shell.cpp`, when `ourgets(buf)` is run, the shell
will start and the user can enter a command. If it begins with an `!`, it
will be run automatically by the system.

Otherwise, `setArgsGiven(buf, arg, types, nArgsMax)` will execute, first
setting all the args and their types to 0, then separating the command itself
from its args, and finally storing the args in an `arg` array while
simultaneously determining if the args are words or numbers (with a placeholder
for the types of the args stored in a `types` array).

Next, `findCmd(buf, types)` executes. It iterates through the commands in
the `cmdTable`, searching for a command whose name and reqired args types
match the name and args types of the previously entered command. If a match
is found, the index of the command in the `cmdTable` is returned.

Finally, `invokeCmd(k, arg)` uses this index and the previously stored args
to call the function corresponding to the correct command in the `cmdTable`
if the args are as they should be and any needed simDisk or cfv exist.