Implemented creation of sym links but could use work
They don't use the sym link flag, instead defaulting to the type of the link's
source. Sym links to files are just completely new files basically with either
a new name or the same name (depending on how the call was made).

Also fixed logic error in flag check for recursive ls and rm