Messing with hard links and soft links
Hard links work but are incomplete because a hard link may be made to a soft link
Soft links are breaking. Paths are not being stored correctly for some reason,
presumably because the wrong block is being retrieved. Rm has been altered
for soft link compatibility, a recursive method to get the iType of the entity
a soft link is pointing to has been created, cd and MvPath need to deal with soft links.
All of this should be tested.